fix: fix slices
