+++ src/components/Layout.test.tsx
@@ -18,12 +18,7 @@ const mockRouter = {
 
 jest.mock('../hooks/custom', () => ({
   useAppDispatch: () => jest.fn(),
-  useAppSelector: () => jest.fn(),
-}));
-jest.mock('../store/reducers/perPageSlice', () => ({
-  perPageSlice: {
-    actions: { newLimit: jest.fn() },
-  },
+  useAppSelector: jest.fn().mockReturnValue({ qty: '20', searchStr: '' }),
 }));
 
 describe('Layout', () => {
