+++ src/components/Pagination.tsx
@@ -1,7 +1,7 @@
 import { useRouter } from 'next/router';
 import { useState } from 'react';
 import { Pokemon } from '../api/types';
-import { useAppDispatch } from '../hooks/custom';
+import { useAppDispatch, useAppSelector } from '../hooks/custom';
 import { perPageSlice } from '../store/reducers/perPageSlice';
 
 export default function Pagination({ pokemons }: { pokemons: Pokemon[] }) {
@@ -9,9 +9,9 @@ export default function Pagination({ pokemons }: { pokemons: Pokemon[] }) {
   const dispatch = useAppDispatch();
   const { newLimit } = perPageSlice.actions;
   const pageId = router.query.pageId;
-  const qty = router.query.qty;
+  const { qty } = useAppSelector((state) => state.perPageReducer);
   const [value, setValue] = useState(qty);
-  const maxPage = Math.ceil(1200 / Number(value));
+  const maxPage = Math.ceil(1200 / Number(qty));
   const nextPage =
     Number(pageId) < maxPage && pokemons.length !== 1
       ? Number(pageId) + 1
@@ -21,7 +21,7 @@ export default function Pagination({ pokemons }: { pokemons: Pokemon[] }) {
   function changeValue(event: React.ChangeEvent<HTMLInputElement>) {
     setValue(event.target.value);
     dispatch(newLimit(event.target.value));
-    router.push(`/page/1?qty=${event.target.value}`);
+    router.push(`/page/1`);
   }
 
   return (
@@ -31,7 +31,7 @@ export default function Pagination({ pokemons }: { pokemons: Pokemon[] }) {
           <button
             disabled={prevPage === 0 || pokemons.length === 1}
             className="navBtn"
-            onClick={() => router.push(`/page/${prevPage}?qty=${value}`)}
+            onClick={() => router.push(`/page/${prevPage}`)}
           >
             prev
           </button>
@@ -39,7 +39,7 @@ export default function Pagination({ pokemons }: { pokemons: Pokemon[] }) {
           <button
             disabled={!nextPage}
             className="navBtn"
-            onClick={() => router.push(`/page/${nextPage}?qty=${value}`)}
+            onClick={() => router.push(`/page/${nextPage}`)}
           >
             next
           </button>
